+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Allow calls to self to fix name clash with argument [#484](https://github.com/bbatsov/rubocop/pull/484)
 * Renamed `SpaceAroundBraces` to `SpaceAroundBlockBraces`.
 * `SpaceAroundBlockBraces` now has a `NoSpaceBeforeBlockParameters` config option to enforce a style for blocks with parameters like `{|foo| puts }`.
+* New cop `LambdaCall` tracks uses of the obscure `lambda.(...)` syntax
 
 ### Bugs fixed